Ported changes from ZDoom revision 1223-1229 and maintenance changes from GZDoom revision 179-180: - fixed: For Chex Quest some CVars were initialized to Heretic's default. - Cleaned up some include dependencies. - More header cleanup. - Changed compilation for g_doom, g_heretic, g_hexen and g_strife folders so that all files are included by a central one instead of compiling each one separately. This speeds up the compilation process by 25% when doing a complete rebuild in Visual C. - Cleaned up more header dependencies. - Moved some more high level sound logic out of FMODSoundRenderer: The rolloff and channel ended callbacks now call functions in s_sound.cpp instead of working on the data itself and GSnd->StopSound has been replaced with S_StopChannel. - Separated low level sound code from all high level dependencies. - fixed an uninitialized variable in sound code.
SVN r1967 (latestzdoom)

+++ src/nodebuild.cpp
@@ -199,7 +199,7 @@ void FNodeBuilder::CreateSubsectorsForReal ()
 		// Convert seg pointers into indices
 		for (unsigned int i = sub.firstline; i < SegList.Size(); ++i)
 		{
-			SegList[i].SegNum = SegList[i].SegPtr - &Segs[0];
+			SegList[i].SegNum = DWORD(SegList[i].SegPtr - &Segs[0]);
 		}
 		Subsectors.Push (sub);
 	}
